requirements.txt
fastapi==0.77.1
pydantic[email]
uvicorn==0.22.0
tenacity==8.0.1
celery[redis]==5.3.1
requests==2.27.1
python-dotenv==1.0.0

swmm-toolkit==0.9.1
swmmio==0.6.2